* Frameworks/GNUmakefile Included XMPPKit in the build process.
* Services/User/Jabber
Deleted xmpp directory. Updated GNUmakefile to link the necessary libraries 
and removed old stuff related to the aggregate project. Updated all import 
statements related to XMPPKit to reference the headers as part of the deployed 
framework and not local headers located in xmpp subdirectory.
Take note the Xcode project is currently broken as a side-effect.

The whole repository builds fine for me and StepChat still runs :-) Extra
testing welcome as usual.

git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Services/User/Jabber@3445 7b3f36aa-e6db-49c1-8b4c-66eac1997e64